🔧 fix: 修复numpy类型导致FastAPI序列化500错误
核心问题：numpy.int64和numpy.float32无法被FastAPI序列化
- tail_vector_manager.py: 转换sample_id和vector_index为Python int类型
- intelligent_tail_filter.py: 转换avg_confidence为Python float类型
- 消除"实现细节泄漏"问题（Linus风格）

错误路径：
numpy.int64 → filter_result.details → FastAPI JSON encoder → TypeError

修复验证：
- 过滤功能正常执行
- 现在返回401认证错误而非500服务器错误 ✅
- API能正确序列化响应数据

+++ app/services/intelligent_tail_filter.py
@@ -167,7 +167,7 @@ class IntelligentTailFilter:
         
         # 调整得分
         if similar_samples:
-            avg_confidence = np.mean([s['similarity'] for s in similar_samples])
+            avg_confidence = float(np.mean([s['similarity'] for s in similar_samples]))  # 转换numpy类型为Python float
             if avg_confidence > 0.8:
                 scores['overall_score'] = min(scores['overall_score'] + 0.2, 1.0)
